board: zondax: imx8mmevk: fix hab event issue
Now, CSF files used to sign the SPL and the FIT image are filled, just
before to sign images, in order to prevent bad offset or size.

And add an IVT offset otherwise authenticate datas at the end of the FIT
image can't be verified, because they aren't fully read.

Signed-off-by: Thomas Perrot <thomas.perrot@bootlin.com>